adding protons and neutrons
added proton and neutrons to NucleusInfo, as well as the Hulthen PDF
parameters - these are set to zero for anything except for deuteron.
Protons are a special case - they only have a radius, which is set to a
small value.  This is used for a radial step function that approximates
a delta function for the proton position